Add a dropdown menu for the admin section
