+++ web-ui/src/main/resources/catalog/js/GnModule.js
@@ -27,6 +27,7 @@
   goog.require('gn_search_form_controller');
   goog.require('gn_search_manager');
   goog.require('gn_utility');
+  goog.require('gn_admin_menu');
 
   /**
    * GnModule just manage angular injection with
@@ -49,7 +50,8 @@
     'gn_map',
     'gn_search_form_controller',
     'gn_needhelp',
-    'gn_alert'
+    'gn_alert',
+    'gn_admin_menu'
   ]);
 
 })();
