+++ web-ui/src/main/resources/catalog/js/admin/AdminController.js
@@ -23,13 +23,14 @@
   goog.require('gn_settings_controller');
   goog.require('gn_standards_controller');
   goog.require('gn_usergroup_controller');
+  goog.require('gn_admin_menu');
 
   var module = angular.module('gn_admin_controller',
       ['gn_dashboard_controller', 'gn_usergroup_controller',
        'gn_admintools_controller', 'gn_settings_controller',
        'gn_adminmetadata_controller', 'gn_classification_controller',
        'gn_harvest_controller', 'gn_standards_controller',
-       'gn_report_controller']);
+       'gn_report_controller','gn_admin_menu']);
 
 
   var tplFolder = '../../catalog/templates/admin/';
@@ -114,48 +115,10 @@
    */
   module.controller('GnAdminController', [
     '$scope', '$http', '$q', '$rootScope', '$route', '$routeParams',
-    'gnUtilityService',
+    'gnUtilityService', 'gnAdminMenu',
     function($scope, $http, $q, $rootScope, $route, $routeParams,
-        gnUtilityService) {
-      /**
-       * Define admin console menu for each type of user
-       */
-      var userAdminMenu = [
-        {name: 'harvesters', route: '#harvest',
-          classes: 'btn-primary', icon: 'fa-cloud-download'},
-        {name: 'statisticsAndStatus', route: '#dashboard',
-          classes: 'btn-success', icon: 'fa-dashboard'},
-        {name: 'reports', route: '#reports',
-          classes: 'btn-success', icon: 'fa-file-text-o'},
-        {name: 'usersAndGroups', route: '#organization',
-          classes: 'btn-default', icon: 'fa-group'}
-
-      ];
-      $scope.menu = {
-        UserAdmin: userAdminMenu,
-        Administrator: [
-          // TODO : create gn classes
-          {name: 'metadatasAndTemplates', route: '#metadata',
-            classes: 'btn-primary', icon: 'fa-archive'},
-          {name: 'harvesters', route: '#harvest', //url: 'harvesting',
-            classes: 'btn-primary', icon: 'fa-cloud-download'},
-          {name: 'statisticsAndStatus', route: '#dashboard',
-            classes: 'btn-success', icon: 'fa-dashboard'},
-          {name: 'reports', route: '#reports',
-            classes: 'btn-success', icon: 'fa-file-text-o'},
-          {name: 'classificationSystems', route: '#classification',
-            classes: 'btn-info', icon: 'fa-tags'},
-          {name: 'standards', route: '#standards',
-            classes: 'btn-info', icon: 'fa-puzzle-piece'},
-          {name: 'usersAndGroups', route: '#organization',
-            classes: 'btn-default', icon: 'fa-group'},
-          {name: 'settings', route: '#settings',
-            classes: 'btn-warning', icon: 'fa-gear'},
-          {name: 'tools', route: '#tools',
-            classes: 'btn-warning', icon: 'fa-medkit'}]
-        // TODO : add other role menu
-      };
-
+        gnUtilityService,gnAdminMenu) {
+      $scope.menu = gnAdminMenu;
       /**
        * Define menu position on the left (nav-stacked)
        * or on top of the page.
