+++ web-ui/src/main/resources/catalog/templates/top-toolbar.html
@@ -73,6 +73,17 @@
           <span class="visible-lg" data-translate="">adminConsole</span>
         </a>
       </li>
+      <li class="dropdown" data-ng-show="user.isReviewerOrMore() && viewMenuAdmin">
+        <a href="" class="dropdown-toggle" data-toggle="dropdown"
+           role="button" aria-expanded="false"><span class="caret"></span></a>
+        <ul class="dropdown-menu" role="menu">
+          <li data-ng-repeat="t in adminMenu">
+            <a data-gn-active-tb-item="admin.console{{t.route}}">
+              <i class="fa {{t.icon}}"/>&nbsp;<span data-translate="">{{t.name | translate}}</span>
+            </a>
+          </li>
+        </ul>
+      </li>
     </ul>
